+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add automatic periodic checking for new unstable versions of lilypond (#7).
+
 ## Version 0.3.8 2016-06-14
 
 - Fix behavior when `-dhelp` switch is specified.