+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 0.3.9 2016-06-27
+
+- Add `--auto-install-deps` lilypond option for automatically installing missing dependencies (#25).
 - Add `--snippet` option to lilypond for creating cropped png images (#24).
 - Add `--dev` option for installing development packages (#23).
 - Raise error in mismatching require version clauses in same file (#12).